requirements.txt

fastapi
uvicorn
pymongo
motor
python-jose[cryptography]
pydantic[email]